pull: Implement a metadata-only pull mode
This just fetches the top level commit object. The intention is to
allow this minimal pre-fetch to be used to determine the size of an
upcoming pull in advance.  Invoking pull again without the
METADATA_ONLY flag fetches the content.

Prior to metadata-only pulls, having the top level object was assumed
to mean you had the whole tree - this is not the case anymore, so we
explicitly store the commits named as "committhin".