libostree: Add ostree_repo_pull_with_options()
We potentially need a lot of argument types for pull.  Rather than
have a C function with tons of arguments, let's use a GVariant a{sv}
as a handy extensible (and immutable) bag of properties.

This is prepratory work for adding an option to pull to traverse
history.

The first additional flag here is inherit-maincontext, which
is used to fix

https://bugzilla.gnome.org/show_bug.cgi?id=737844