refs: added "ostree refs --create" functionality and associated unit tests
Added the ability to create a ref (much like a git tag) for
an existing commit through "ostree refs EXISTING --create=NEWREF".
Previously the only way to create a new ref was by creating a new commit,
but refs --create allows multiple refs to point to the same commit.

The command will fail if:
1. None/more than one existing ref is specified
2. The specified EXISTING tag does not exist, or was not specified
3. The specified NEWREF already exists, or is the name of a folder
The above conditions are tested and verified through additional
unit tests in tests-ref.sh